// system-design-mastery/saga-pattern-orchestration/apps/order-orchestration/src/enum/order.enum.ts
export enum OrderSagaStatus {
  STARTED = 'STARTED',
  PAYMENT_PROCESSING = 'PAYMENT_PROCESSING',
  PAYMENT_COMPLETED = 'PAYMENT_COMPLETED',
  INVENTORY_PROCESSING = 'INVENTORY_PROCESSING',
  INVENTORY_COMPLETED = 'INVENTORY_COMPLETED',
  COMPENSATING = 'COMPENSATING',
  COMPENSATED = 'COMPENSATED',
  COMPLETED = 'COMPLETED',
  FAILED = 'FAILED',
}